refactor app.rs
