+++ src/main.rs
@@ -10,6 +10,7 @@
     clippy::new_without_default
 )]
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
+
 pub mod app;
 pub mod audio;
 mod cacher;
